+++ extensions/Bilibili/src/apis/logout.ts
@@ -1,21 +1,18 @@
 import { API } from "./api";
-
-import got from "got";
+import { getJson } from "./request";
 import { Cache } from "@raycast/api";
 
 export async function logout() {
   const cache = new Cache();
   const cookie = cache.get("cookie");
 
-  const response: Bilibili.LogoutResponse = await got
-    .post(API.logout(), {
-      headers: {
-        cookie,
-        "Content-Type": "application/x-www-form-urlencoded",
-      },
-      body: `biliCSRF=${cookie?.match(/bili_jct=(.*?);/)?.[1]}`,
-    })
-    .json();
+  const response = await getJson<Bilibili.LogoutResponse>(API.logout(), {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/x-www-form-urlencoded",
+    },
+    body: `biliCSRF=${cookie?.match(/bili_jct=(.*?);/)?.[1]}`,
+  });
 
   if (response.code !== 0) throw new Error(response.message);