+++ extensions/Bilibili/src/apis/postHeartbeat.ts
@@ -1,21 +1,18 @@
 import { API } from "./api";
-
-import got from "got";
+import { getJson } from "./request";
 import { Cache } from "@raycast/api";
 
 export async function postHeartbeat(aid: number, cid: number) {
   const cache = new Cache();
   const cookie = cache.get("cookie");
 
-  const response: Bilibili.BaseResponse = await got
-    .post(API.heartbeat(), {
-      headers: {
-        cookie,
-        "Content-Type": "application/x-www-form-urlencoded",
-      },
-      body: `played_time=1&aid=${aid}&cid=${cid}&csrf=${cookie?.match(/bili_jct=(.*?);/)?.[1]}`,
-    })
-    .json();
+  const response = await getJson<Bilibili.BaseResponse>(API.heartbeat(), {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/x-www-form-urlencoded",
+    },
+    body: `played_time=1&aid=${aid}&cid=${cid}&csrf=${cookie?.match(/bili_jct=(.*?);/)?.[1]}`,
+  });
 
   if (response.code !== 0) throw new Error(response.message);