+++ extensions/Bilibili/src/components/index.ts
@@ -1,3 +1,4 @@
 export * from "./video";
+export * from "./videoGrid";
 export * from "./post";
 export * from "./checkLogin";